Added RATE control for the UAV
